Make `lookup_action` a function, rather than a bug method
