+++ jbi/runner.py
@@ -4,6 +4,7 @@ Execute actions from Webhook requests
 import inspect
 import itertools
 import logging
+import re
 from typing import Optional
 
 from statsd.defaults.env import statsd
@@ -13,6 +14,7 @@ from jbi import steps as steps_module
 from jbi.environment import get_settings
 from jbi.errors import ActionNotFoundError, IgnoreInvalidRequestError
 from jbi.models import (
+    Action,
     ActionContext,
     ActionParams,
     Actions,
@@ -47,6 +49,24 @@ def groups2operation(steps: ActionSteps):
     except KeyError as err:
         raise ValueError(f"Unsupported entry in `steps`: {err}") from err
     return by_operation
+
+
+def lookup_action(bug: bugzilla.BugzillaBug, actions: Actions) -> Action:
+    """
+    Find first matching action from bug's whiteboard field.
+
+    Tags are strings between brackets and can have prefixes/suffixes
+    using dashes (eg. ``[project]``, ``[project-moco]``, ``[project-moco-sprint1]``).
+    """
+
+    if bug.whiteboard:
+        for tag, action in actions.by_tag.items():
+            # [tag-word], [tag-], [tag], but not [word-tag] or [tagword]
+            search_string = r"\[" + tag + r"(-[^\]]*)*\]"
+            if re.search(search_string, bug.whiteboard, flags=re.IGNORECASE):
+                return action
+
+    raise ActionNotFoundError(", ".join(actions.by_tag.keys()))
 
 
 class Executor:
@@ -174,7 +194,7 @@ def execute_action(
 
         runner_context = runner_context.update(bug=bug)
         try:
-            action = bug.lookup_action(actions)
+            action = lookup_action(bug, actions)
         except ActionNotFoundError as err:
             raise IgnoreInvalidRequestError(
                 f"no bug whiteboard matching action tags: {err}"
