+++ tests/unit/test_models.py
@@ -1,7 +1,6 @@
 import pydantic
 import pytest
 
-from jbi.errors import ActionNotFoundError
 from jbi.models import ActionParams, Actions, ActionSteps
 
 
@@ -105,54 +104,6 @@ def test_extract_see_also(see_also, expected, bug_factory):
 def test_product_component(product, component, expected, bug_factory):
     bug = bug_factory(product=product, component=component)
     assert bug.product_component == expected
-
-
-@pytest.mark.parametrize(
-    "whiteboard",
-    [
-        "[DevTest]",
-        "[DevTest-]",
-        "[DevTest-test]",
-        "[DevTest-test-foo]",
-        "[example][DevTest]",
-        "[DevTest][example]",
-        "[example][DevTest][example]",
-    ],
-)
-def test_lookup_action_found(whiteboard, actions, bug_factory):
-    bug = bug_factory(id=1234, whiteboard=whiteboard)
-    action = bug.lookup_action(actions)
-    assert action.whiteboard_tag == "devtest"
-    assert "test config" in action.description
-
-
-@pytest.mark.parametrize(
-    "whiteboard",
-    [
-        "DevTest",
-        "[-DevTest-]",
-        "[-DevTest]",
-        "[test-DevTest]",
-        "[foo-DevTest-bar]",
-        "[foo-bar-DevTest-foo-bar]",
-        "foo DevTest",
-        "DevTest bar",
-        "foo DevTest bar",
-        "[fooDevTest]",
-        "[foo DevTest]",
-        "[DevTestbar]",
-        "[DevTest bar]",
-        "[fooDevTestbar]",
-        "[fooDevTest-bar]",
-        "[foo-DevTestbar]",
-        "[foo] devtest [bar]",
-    ],
-)
-def test_lookup_action_not_found(whiteboard, actions, bug_factory):
-    bug = bug_factory(id=1234, whiteboard=whiteboard)
-    with pytest.raises(ActionNotFoundError) as exc_info:
-        bug.lookup_action(actions)
-    assert str(exc_info.value) == "devtest"
 
 
 def test_payload_empty_changes_list(webhook_event_factory):
